Use WebAssembly backend for the minimax web worker
[More info](https://www.scala-js.org/doc/project/webassembly.html)

+++ frontend/src/main/scala/be/doeraene/communication/WorkerAPI.scala
@@ -2,7 +2,7 @@ package be.doeraene.communication
 
 import be.doeraene.workers.*
 import org.scalajs.dom
-import org.scalajs.dom.Worker
+import org.scalajs.dom.{Worker, WorkerOptions, WorkerType}
 import scala.concurrent.{Future, Promise}
 import be.doeraene.globals.webWorkerPath
 
@@ -16,29 +16,40 @@ object WorkerAPI:
   def makeWorkerCompute[T <: WorkerProtocol](message: T)(using ClassTag[message.Response]): Future[message.Response] = {
     val promise = Promise[message.Response]()
 
-    val worker = Worker(webWorkerPath)
+    val workerOptions = new WorkerOptions {
+      `type` = WorkerType.module
+    }
+    val worker = Worker(webWorkerPath, workerOptions)
     worker.onerror = (event: dom.ErrorEvent) => {
       worker.terminate()
       dom.console.error(event)
       promise.failure(ProtocolException.DomErrorWrapper(event))
     }
 
-    worker.onmessage = (event: dom.MessageEvent) => {
-      worker.terminate()
-      val responseOrFailure = for {
-        data <- Option(event.data).toRight(new ProtocolException.ReceivedDataWasNull)
-        stringData <- data match {
-          case str: String => Right(str)
-          case other       => Left(new ProtocolException.ReceivedNonStringData(data))
-        }
-        decodedData <- decode[WorkerProtocol](stringData).swap.map(new ProtocolException.DecodingError(_)).swap
-        response    <- message.collectResponse(decodedData)
-      } yield response
-
-      promise.complete(responseOrFailure.toTry)
-    }
     val toSend = (message: WorkerProtocol).asJson.noSpaces
-    worker.postMessage(toSend)
+
+    // The worker sends WorkerProtocol.readySignal as soon as it is ready to receive its first
+    // message; see be.doeraene.EntryPoint for why we must wait for it before posting anything.
+    var workerIsReady = false
+
+    worker.onmessage = (event: dom.MessageEvent) =>
+      if !workerIsReady && event.data == WorkerProtocol.readySignal then {
+        workerIsReady = true
+        worker.postMessage(toSend)
+      } else {
+        worker.terminate()
+        val responseOrFailure = for {
+          data <- Option(event.data).toRight(new ProtocolException.ReceivedDataWasNull)
+          stringData <- data match {
+            case str: String => Right(str)
+            case other       => Left(new ProtocolException.ReceivedNonStringData(data))
+          }
+          decodedData <- decode[WorkerProtocol](stringData).swap.map(new ProtocolException.DecodingError(_)).swap
+          response    <- message.collectResponse(decodedData)
+        } yield response
+
+        promise.complete(responseOrFailure.toTry)
+      }
 
     promise.future
   }
